+++ examples/ex20_third-medium-contact.py
@@ -8,11 +8,11 @@ Third Medium Contact
 
    * define multiple solid bodies and create a top-level field
 
-   * create an animation of the deformed solid
+   * plot the deformed solid
 
 This contact method uses a third medium for two solid contact bodies with a a Hessian-
-based regularization [2]_. First, let's create sub meshes with quad cells for the solid
-body. All sub meshes are merged by stacking the meshes of the
+based regularization [2]_. First, let's create sub meshes with bi-quadratic quad cells
+for the solid body. All sub meshes are merged by stacking the meshes of the
 :class:`mesh container <felupe.MeshContainer>` into a :class:`mesh <felupe.Mesh>`.
 """
 
@@ -32,6 +32,7 @@ c = fem.Rectangle(a=(t, 0), b=(L, t), n=(nL, nt))
 d = c.translate(move=H - t, axis=1)
 e = fem.Rectangle(a=(0, t), b=(t, H - t), n=(nt, nH))
 body = fem.MeshContainer([a, b, c, d, e], merge=True).stack()
+body = body.add_midpoints_edges().add_midpoints_faces()
 
 # %%
 # This is repeated for the medium, where two meshes are added to a mesh container. Extra
@@ -39,6 +40,7 @@ body = fem.MeshContainer([a, b, c, d, e], merge=True).stack()
 f = fem.Rectangle(a=(t, t), b=(L, H - t), n=(nL, nH))
 g = fem.Rectangle(a=(L, 0), b=(L + t / (nt - 1), H), n=(2, 2 * (nt - 1) + nH))
 medium = fem.MeshContainer([f, g], merge=True).stack()
+medium = medium.add_midpoints_edges().add_midpoints_faces()
 
 # %%
 # Both stacked meshes are added to a top-level mesh container. For each mesh of the
@@ -48,13 +50,16 @@ medium = fem.MeshContainer([f, g], merge=True).stack()
 # down material parameters.
 G = 5 / 14
 K = 5 / 3
-kr = 1e-6
-gamma0 = 1e-6
+kr = 5e-6
+gamma0 = 5e-6
 
 container = fem.MeshContainer([body, medium], merge=True, decimals=6)
-regions = [fem.RegionQuad(m, hess=True) for m in container.meshes]
+regions = [fem.RegionBiQuadraticQuad(m, hess=True) for m in container.meshes]
 fields = [fem.FieldPlaneStrain(r, dim=2).as_container() for r in regions]
-umats = [fem.NeoHooke(mu=G, bulk=K), fem.NeoHooke(mu=G * gamma0, bulk=K * gamma0)]
+umats = [
+    fem.NeoHookeCompressible(mu=G, lmbda=K),
+    fem.NeoHookeCompressible(mu=G * gamma0, lmbda=K * gamma0),
+]
 solids = [fem.SolidBody(umat, f) for umat, f in zip(umats, fields)]
 
 # %%
@@ -62,37 +67,56 @@ solids = [fem.SolidBody(umat, f) for umat, f in zip(umats, fields)]
 # unknowns and is required for the selection of prescribed degrees of freedom as well
 # as for Newton's method. The left end edge is fixed and the vertical displacement is
 # prescribed on the outermost top-right point.
-region = fem.RegionQuad(container.stack())
+region = fem.RegionBiQuadraticQuad(container.stack())
 field = fem.FieldPlaneStrain(region, dim=2).as_container()
 
-bounds = dict(
+bounds = fem.BoundaryDict(
     fixed=fem.Boundary(field[0], fx=0),
     move=fem.Boundary(field[0], fx=L, fy=H, mode="and", skip=(1, 0)),
 )
 
 # %%
-# The so-called HuHu-regularization is created by two weak-:func:`forms <felupe.Form>`,
-# which are derived from the regularization potential, see Eq.
-# :eq:`huhu-regularization` [3]_.
+# The so-called HuHu-LuLu-regularization is created by two weak-
+# :func:`forms <felupe.Form>`, which are derived from the regularization
+# potential, see Eq. :eq:`huhu-regularization` [3]_.
 #
 # .. math::
-#    :label: huhu-regularization
+#    :label: huhu-lulu-regularization
 #
-#    \Pi = \frac{1}{2} \int_V
-#        \mathbb{H}(\boldsymbol{u})~\vdots~\mathbb{H}(\boldsymbol{u})~dV
+#    \Psi(\boldsymbol{u}) = W(\boldsymbol{u}) +
+#        \mathbb{H}(\boldsymbol{u})~\vdots~\mathbb{H}(\boldsymbol{u})
+#        - \frac{1}{\text{tr}(\boldsymbol{1})}
+#          \mathbb{L}(\boldsymbol{u}) \cdot \mathbb{L}(\boldsymbol{u})
 #
-from felupe.math import dddot, hess
+from felupe.math import dot, dddot, hess
+
+laplacian = lambda x: np.einsum("ijj...->i...", x)
 
 
 @fem.Form(v=fields[1], u=fields[1], kwargs={"kr": 1.0})
 def bilinearform():
-    return [lambda v, u, kr: kr * dddot(hess(u), hess(v))]
+    def a(v, u, kr):
+        Hu = hess(u)
+        Hv = hess(v)
+        Lu = laplacian(Hu)
+        Lv = laplacian(Hv)
+        return kr * (dddot(Hu, Hv) - dot(Lu, Lv, mode=(1, 1)) / 2)
+
+    return [a]
 
 
 @fem.Form(v=fields[1], kwargs={"kr": 1.0})
 def linearform():
     u = fields[1][0]
-    return [lambda v, kr: kr * dddot(hess(v), hess(u)[:2, :2, :2])]
+
+    def L(v, kr):
+        Hu = hess(u)[:2, :2, :2]
+        Hv = hess(v)
+        Lu = laplacian(Hu)
+        Lv = laplacian(Hv)
+        return kr * (dddot(Hu, Hv) - dot(Lu, Lv, mode=(1, 1)) / 2)
+
+    return [L]
 
 
 regularization = fem.FormItem(
@@ -102,34 +126,27 @@ regularization = fem.FormItem(
 )
 
 # %%
-# The prescribed displacement is ramped up to the maximum value and released until zero.
-move = fem.math.linsteps([0, 1, 0], num=30)
+# The prescribed displacement is ramped up to the maximum value.
+move = fem.math.linsteps([0, 1], num=20)
 step = fem.Step(
     items=[*solids, regularization],
-    ramp={bounds["move"]: -0.4 * move * L},
+    ramp={bounds["move"]: -0.62 * move * L},
     boundaries=bounds,
 )
 
 # %%
-# The deformation is captured in a GIF-file. The top-level field has to be passed as the
-# ``x0``-argument for :func:`Newton's method <felupe.newtonrhapson>`, which is called on
-# evaluation. After all frames are recorded, it is important to ``close()`` the plotter.
-plotter = container.plot(colors=["grey", "white"], line_width=3, off_screen=True)
-plotter.open_gif("third-medium-contact.gif", fps=5)
+# The top-level field has to be passed as the ``x0``-argument of the job.
+# The deformation configuration is plotted.
+job = fem.Job([step]).evaluate(x0=field)
 
-
-def record(stepnumber, substepnumber, substep, plotter):
-    "Update the mesh of the plotter and write a frame."
-    if substepnumber in np.arange(len(move), step=5):
-        for m in plotter.meshes:
-            m.points[:, :2] = region.mesh.points + field[0].values
-        plotter.write_frame()
-
-
-job = fem.Job([step], callback=record, plotter=plotter)
-job.evaluate(x0=field)
-
-plotter.close()
+plotter = solids[1].plot(nonlinear_subdivision=2, opacity=0.2, edge_color="grey")
+plotter = solids[0].plot(
+    "Displacement",
+    component=None,
+    nonlinear_subdivision=2,
+    plotter=plotter,
+)
+plotter.show()
 
 # %%
 # References