+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file. The format 
 ## [9.4.1] - 2025-09-27
 
 ### Added
--Add the pyOpenSci peer reviewed badge.
+- Add the pyOpenSci peer reviewed badge.
 
 ## [9.4.0] - 2025-09-25