+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Add a JOSS badge to README.md and change CITATION.cff to JOSS.
-- Add the keyword `Job.evaluate(tqdm="tqdm")` and `newtonrhapson(tqdm="tqdm")` with additional options `"auto"` and `"notebook"`. The `tqdm`-keyword allows to switch the tqdm-backend manually. Note that in a Jupyter console, a progress bar from `tqdm.auto` does not update. 
+- Add the keyword `Job.evaluate(tqdm="tqdm")` and `newtonrhapson(tqdm="tqdm")` with additional options `"auto"` and `"notebook"`. The `tqdm`-keyword allows to switch the tqdm-backend manually. Note that in a Jupyter console, a progress bar from `tqdm.auto` does not update.
+- Add the `hessian()` method to the `ArbitraryOrderLagrange` element and also for `BiQuadraticQuad` and `TriQuadraticHexahedron`.
 
 ### Changed
 - Change the recommended citation from Zenodo to JOSS in README.md.